+++ src/cli/api.rs
@@ -1,6 +1,7 @@
 use crate::api::BrokerSearchResponse::BadRequestResponse;
 use bgpkit_broker::{BrokerItem, LocalBrokerDb, DEFAULT_PAGE_SIZE};
 use chrono::{Duration, NaiveDateTime};
+use clap::Args;
 use poem::listener::TcpListener;
 use poem::middleware::AddData;
 use poem::web::Data;
@@ -10,6 +11,41 @@ use serde::{Deserialize, Serialize};
 use std::str::FromStr;
 
 struct BrokerAPI;
+
+#[derive(Object, Args, Debug, Serialize, Deserialize)]
+pub struct BrokerSearchQuery {
+    /// Start timestamp
+    #[clap(short = 't', long)]
+    pub ts_start: Option<String>,
+
+    /// End timestamp
+    #[clap(short = 'T', long)]
+    pub ts_end: Option<String>,
+
+    /// duration before `ts_end` or after `ts_start`
+    #[clap(short, long)]
+    pub duration: Option<String>,
+
+    /// filter by route collector projects, i.e. `route-views` or `riperis`
+    #[clap(short, long)]
+    pub project: Option<String>,
+
+    /// filter by collector IDs, e.g. 'rrc00', 'route-views2. use comma to separate multiple collectors
+    #[clap(short, long)]
+    pub collectors: Option<String>,
+
+    /// filter by data types, i.e. 'update', 'rib'.
+    #[clap(short, long)]
+    pub data_type: Option<String>,
+
+    /// page number
+    #[clap(long)]
+    pub page: Option<usize>,
+
+    /// page size
+    #[clap(long)]
+    pub page_size: Option<usize>,
+}
 
 #[derive(Object, Serialize, Deserialize, Clone, Debug)]
 struct BrokerSearchResult {
@@ -192,12 +228,12 @@ async fn search(database: Data<&LocalBrokerDb>) -> Json<serde_json::Value> {
 }
 
 pub async fn start_api_service(database: LocalBrokerDb, socket_addr: &str) -> std::io::Result<()> {
-    let api_service = OpenApiService::new(BrokerAPI, "BGPKIT Broker", "3.0.0").server("/api");
+    let api_service = OpenApiService::new(BrokerAPI, "BGPKIT Broker", "3.0.0").server("/");
     let ui = api_service.swagger_ui();
 
     let route = Route::new()
-        .nest("/api", api_service)
-        .nest("/", ui)
+        .nest("/", api_service)
+        .nest("/docs", ui)
         .with(AddData::new(database));
 
     Server::new(TcpListener::bind(socket_addr)).run(route).await
