--- src/cli/search.rs
@@ -1,38 +1,0 @@
-use clap::Args;
-use poem_openapi::Object;
-use serde::{Deserialize, Serialize};
-
-#[derive(Object, Args, Debug, Serialize, Deserialize)]
-pub struct BrokerSearchQuery {
-    /// Start timestamp
-    #[clap(short = 't', long)]
-    pub ts_start: Option<String>,
-
-    /// End timestamp
-    #[clap(short = 'T', long)]
-    pub ts_end: Option<String>,
-
-    /// duration before `ts_end` or after `ts_start`
-    #[clap(short, long)]
-    pub duration: Option<String>,
-
-    /// filter by route collector projects, i.e. `route-views` or `riperis`
-    #[clap(short, long)]
-    pub project: Option<String>,
-
-    /// filter by collector IDs, e.g. 'rrc00', 'route-views2. use comma to separate multiple collectors
-    #[clap(short, long)]
-    pub collectors: Option<String>,
-
-    /// filter by data types, i.e. 'update', 'rib'.
-    #[clap(short, long)]
-    pub data_type: Option<String>,
-
-    /// page number
-    #[clap(long)]
-    pub page: Option<usize>,
-
-    /// page size
-    #[clap(long)]
-    pub page_size: Option<usize>,
-}
